backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
transformers==4.11.3
torch==1.9.1
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5
# 阿里云千问API依赖已包含在requests中